fix: 채팅 생성과 가입 후 소켓 연결 로직 적용 (#278)

+++ pennyway-client-iOS/pennyway-client-iOS/Api/Domain/ChatDomain/Dto/Response/Chat/GetJoinedChatRoomsResponseDto.swift
@@ -16,6 +16,6 @@ struct GetJoinedChatRoomsResponseDto: Decodable {
     }
 
     struct ChatRoomDto: Decodable {
-        let chatRoomIds: [Int]
+        let chatRoomIds: [Int64]
     }
 }
